+++ MudBlazor.Extensions/wwwroot/js/utils/dialogExt/baseDialogExtensionHandler.js
@@ -14,8 +14,24 @@
 
     order = 99;
 
-    raiseDialogEvent(eventName) {
-        this.dotNetService.invokeMethodAsync('PublishEvent', eventName, this.dialog.id, this.dotNet, this.dialog.getBoundingClientRect());                  
+    raiseDialogEvent(eventName) {        
+        // Get viewport dimensions
+        var windowHeight = window.innerHeight || document.documentElement.clientHeight;
+        var windowWidth = window.innerWidth || document.documentElement.clientWidth;
+
+        // Get scroll positions
+        var scrollX = window.pageXOffset || document.documentElement.scrollLeft;
+        var scrollY = window.pageYOffset || document.documentElement.scrollTop;
+
+        // Extend the rect object with new properties
+        var extendedRect = {
+            windowHeight: windowHeight,
+            windowWidth: windowWidth,
+            scrollX: scrollX,
+            scrollY: scrollY
+        };
+        const rect = Object.assign(extendedRect, JSON.parse(JSON.stringify(this.dialog.getBoundingClientRect())));        
+        this.dotNetService.invokeMethodAsync('PublishEvent', eventName, this.dialog.id, this.dotNet, rect);
     }
 
     getAnimationDuration() {